chore: move bottom calculating codes to botton section (#373)
