+++ src/pages/Seller/SellerOpenConsult.tsx
@@ -5,7 +5,7 @@ import IsSendPopup from 'components/Seller/SellerOpenConsult/IsSendPopup';
 import MainQuestionSection from 'components/Seller/SellerOpenConsult/MainQuestionSection';
 import NoConsultSection from 'components/Seller/SellerOpenConsult/NoConsultGraphic';
 import OpenConsultHeader from 'components/Seller/SellerOpenConsult/OpenConsultHeader';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { APP_HEADER_HEIGHT } from 'styles/AppStyle';
@@ -23,44 +23,11 @@ export const OPEN_CONSULT_FOOTER_ID = 'bottom-section-wrapper';
 
 function SellerOpenConsult() {
   const { consultid } = useParams();
+
   const isSendPopupOpen = useRecoilValue(isSendPopupOpenState);
+
   const [isReplying, setIsReplying] = useState(false);
   const [text, setText] = useState<string>('');
-
-  /** In order to solve the issue of the IOS keyboard forcibly pushing up the document to create additional space, the bottom value of the bottom section component is managed as a state when the keyboard is opened.
-   ref: https://stackoverflow.com/questions/43833049/how-to-make-fixed-content-go-above-ios-keyboard */
-  const [bottom, setBottom] = useState(0);
-
-  useEffect(() => {
-    const updateBottom = () => {
-      if (visualViewport) {
-        const newBottom =
-          visualViewport.height < window.innerHeight
-            ? window.innerHeight - visualViewport.height
-            : 0;
-
-        setBottom(newBottom);
-      }
-    };
-
-    const handleResize = () => {
-      if (
-        /iPhone|iPad|iPod/.test(window.navigator.userAgent) &&
-        visualViewport
-      ) {
-        updateBottom();
-      }
-    };
-
-    if (visualViewport) {
-      visualViewport.addEventListener('resize', handleResize);
-    }
-    return () => {
-      if (visualViewport) {
-        visualViewport.removeEventListener('resize', handleResize);
-      }
-    };
-  }, []);
 
   //
   //
@@ -69,7 +36,7 @@ function SellerOpenConsult() {
   return (
     <>
       <div
-        style={{ marginTop: APP_HEADER_HEIGHT }}
+        style={{ paddingTop: APP_HEADER_HEIGHT }}
         onClick={() => {
           setIsReplying(false);
         }}
@@ -78,14 +45,10 @@ function SellerOpenConsult() {
         {consultid === 'all-adopted' ? (
           <NoConsultSection />
         ) : (
-          <div
-            //FIXME: do we need this?
-            className="scroll-container"
-            style={{ height: 'calc(100vh - 10rem)', overflow: 'scroll' }}
-          >
+          <>
             <MainQuestionSection />
             <CommentListSection />
-          </div>
+          </>
         )}
       </div>
       {consultid !== 'all-adopted' && (
@@ -94,7 +57,6 @@ function SellerOpenConsult() {
           setIsReplying={setIsReplying}
           text={text}
           setText={setText}
-          bottom={bottom}
         />
       )}
